getLogDir()
