+++ src/momentOfTruthUtils.ts
@@ -2,6 +2,8 @@
 // Licensed under the MIT License. See License.txt in the project root for license information.
 
 import * as stringMap from '@ts-common/string-map'
+import * as path from 'path'
+import * as fs from 'fs'
 
 export type Issue = {
   readonly type?: string
@@ -31,4 +33,17 @@ export type FinalResult = {
   readonly pullRequest: unknown,
   readonly repositoryUrl: unknown,
   readonly files: stringMap.MutableStringMap<File>
+}
+
+// Creates and returns path to the logging directory
+export function getLogDir() {
+  let logDir = path.resolve('output');
+  if (!fs.existsSync(logDir)) {
+      try {
+          fs.mkdirSync(logDir);
+      } catch (e) {
+          if (e.code !== 'EEXIST') throw e;
+      }
+  }
+  return logDir;
 }
